# enrichment-coordinator-service/docs/api.yaml
swagger: '2.0'
info:
  description: This page lists all the rest apis for the service.
  version: '1.0'
  title: Enrichment Data service
host: 'localhost:8081'
basePath: /
...
paths:
  /A1-EI/v1/eitypes:
    get:
      tags:
        - A1-E Enrichment Data Consumer API
      summary: EI type identifiers
      operationId: getEiTypeIdentifiersUsingGET
      produces:
        - application/json
      responses:
        '200':
          description: EI type identifiers
          schema:
            type: array
            items:
              type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/A1-EI/v1/eitypes/{eiTypeId}':
    get:
      tags:
        - A1-E Enrichment Data Consumer API
      summary: Individual EI type
      operationId: getEiTypeUsingGET
      produces:
        - application/json
      parameters:
        - name: eiTypeId
          in: path
          description: eiTypeId
          required: true
          type: string
      responses:
        '200':
          description: EI type
          schema:
            $ref: '#/definitions/ei_type_info'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Enrichment Information type is not found
          schema:
            $ref: '#/definitions/error_information'
      deprecated: false
...
  /ei-producer/v1/eiproducers:
    get:
      tags:
        - Enrichment Data Producer API
      summary: EI producer identifiers
      operationId: getEiProducerIdentifiersUsingGET
      produces:
        - application/json
      responses:
        '200':
          description: EI producer identifiers
          schema:
            type: array
            items:
              type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/ei-producer/v1/eiproducers/{eiProducerId}':
    get:
      tags:
        - Enrichment Data Producer API
      summary: Individual EI producer
      operationId: getEiProducerUsingGET
      produces:
        - application/json
      parameters:
        - name: eiProducerId
          in: path
          description: eiProducerId
          required: true
          type: string
      responses:
        '200':
          description: EI Jobs
          schema:
            $ref: '#/definitions/producer_registration_info'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Enrichment Information producer is not found
          schema:
            $ref: '#/definitions/error_information'
      deprecated: false
    put:
      tags:
        - Enrichment Data Producer API
      summary: Individual EI producer
      operationId: putEiProducerUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: eiProducerId
          in: path
          description: eiProducerId
          required: true
          type: string
        - in: body
          name: registrationInfo
          description: registrationInfo
          required: true
          schema:
            $ref: '#/definitions/producer_registration_info'
      responses:
        '200':
          description: Producer updated
          schema:
            type: object
        '201':
          description: Producer created
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    delete:
      tags:
        - Enrichment Data Producer API
      summary: Individual EI producer
      operationId: deleteEiProducerUsingDELETE
      produces:
        - application/json
      parameters:
        - name: eiProducerId
          in: path
          description: eiProducerId
          required: true
          type: string
      responses:
        '200':
          description: Not used
          schema:
            type: object
        '204':
          description: Producer deleted
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Producer is not found
          schema:
            $ref: '#/definitions/error_information'
      deprecated: false
  '/ei-producer/v1/eiproducers/{eiProducerId}/eijobs':
    get:
      tags:
        - Enrichment Data Producer API
      summary: EI job definitions
      description: EI job definitions for one EI producer
      operationId: getEiProducerJobsUsingGET
      produces:
        - application/json
      parameters:
        - name: eiProducerId
          in: path
          description: eiProducerId
          required: true
          type: string
      responses:
        '200':
          description: EI jobs
          schema:
            type: array
            items:
              $ref: '#/definitions/producer_ei_job_request'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Enrichment Information producer is not found
          schema:
            $ref: '#/definitions/error_information'
      deprecated: false
  /ei-producer/v1/eitypes:
    get:
      tags:
        - Enrichment Data Producer API
      summary: EI type identifiers
      operationId: getEiTypeIdentifiersUsingGET_1
      produces:
        - application/json
      responses:
        '200':
          description: EI type identifiers
          schema:
            type: array
            items:
              type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/ei-producer/v1/eitypes/{eiTypeId}':
    get:
      tags:
        - Enrichment Data Producer API
      summary: Individual EI Type
      operationId: getEiTypeUsingGET_1
      produces:
        - application/json
      parameters:
        - name: eiTypeId
          in: path
          description: eiTypeId
          required: true
          type: string
      responses:
        '200':
          description: EI type
          schema:
            $ref: '#/definitions/producer_ei_type_info'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Enrichment Information type is not found
          schema:
            $ref: '#/definitions/error_information'
      deprecated: false
...
definitions:
...
  ei_type_info:
    type: object
    properties:
      job_data_schema:
        type: object
        description: Json schema for the job data
    title: ei_type_info
    description: Information for an EI type
  error_information:
    type: object
    properties:
      detail:
        type: string
        example: EI job type not found
        description: A human-readable explanation specific to this occurrence of the problem.
      status:
        type: integer
        format: int32
        example: 404
        description: The HTTP status code generated by the origin server for this occurrence of the problem.
    title: error_information
    description: 'Problem as defined in https://tools.ietf.org/html/rfc7807'
  producer_ei_job_request:
    type: object
    required:
      - ei_job_identity
    properties:
      ei_job_data:
        type: object
        description: Json for the job data
      ei_job_identity:
        type: string
        description: Idenitity of the EI job
      ei_type_identity:
        type: string
        description: Type idenitity for the job
    title: producer_ei_job_request
    description: The body of the EI producer callbacks for EI job creation and deletion
  producer_ei_type_info:
    type: object
    properties:
      ei_job_data_schema:
        type: object
        description: Json schema for the job data
      ei_producer_ids:
        type: array
        description: Registered producers
        items:
          type: string
    title: producer_ei_type_info
    description: Information for an EI type
  producer_ei_type_registration_info:
    type: object
    required:
      - ei_type_identity
    properties:
      ei_job_data_schema:
        type: object
        description: Json schema for the job data
      ei_type_identity:
        type: string
        description: EI type identity
    title: producer_ei_type_registration_info
    description: Information for an EI type
  producer_registration_info:
    type: object
    required:
      - ei_job_creation_callback_url
      - ei_job_deletion_callback_url
      - supported_ei_types
    properties:
      ei_job_creation_callback_url:
        type: string
        description: callback for job creation
      ei_job_deletion_callback_url:
        type: string
        description: callback for job deletion
      supported_ei_types:
        type: array
        description: Supported EI types
        items:
          $ref: '#/definitions/producer_ei_type_registration_info'
    title: producer_registration_info
    description: Information for an EI producer